livekit-agent/requirements.txt
livekit-agents>=1.2.0
livekit-plugins-xai
livekit-plugins-noise-cancellation
livekit-plugins-silero
python-dotenv
